requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt<4
python-jose[cryptography]
pydantic
pydantic-settings
